mcqs_generation_app/requirements.txt
# Flask web server
Flask==3.0.3

# LangChain core + integrations
langchain-core==0.1.54
langchain==0.2.3
langchain-community==0.2.3
langchainhub==0.1.15

# PDF loader
PyMuPDF==1.24.2  # aka fitz

# Pinecone vector DB
pinecone-client==3.2.2

# Embeddings
sentence-transformers==2.7.0
huggingface-hub==0.23.2

# LLM support (Groq / OpenAI)
openai==1.30.1
httpx==0.27.0  # for async support
langchain-groq==0.1.3  # if using Groq-compatible LLMs

# Environment variable loading
python-dotenv==1.0.1

# Data validation / JSON parsing
pydantic==2.7.1

# Utilities
requests==2.31.0
uuid==1.30
